% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregateDistanceMatrix.R
\name{AggregateDistanceMatrix}
\alias{AggregateDistanceMatrix}
\title{Decreases matrix resolution.}
\usage{

}
\arguments{
\item{distance.matrix}{A distance matrix produced by \emph{DistanceMatrix}.}
}
\description{
Decreases the resolution of a distance matrix. It outputs the 10th percentile or the minimum of the values of the original cells inside of every new coarser cell.
}
\author{
Blas Benito <blasbenito@gmail.com>
}

