% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeDistances.R
\name{ComputeDistances}
\alias{ComputeDistances}
\title{Computes values of a random walk.}
\usage{

}
\arguments{
\item{distance.matrix}{Numeric distance matrix created by \emph{DistanceMatrix}.}

\item{random.walk}{Data frame with a random walk created by \emph{GenerateRandomWalk}.}
}
\value{
Data frame, the input random walk with a new column named \emph{distance}
}
\description{
description
}
\author{
Blas Benito <blasbenito@gmail.com>
}

