% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputePsi.R
\name{ComputePsi}
\alias{ComputePsi}
\title{Sequence Slotting}
\usage{

}
\arguments{
\item{sequences}{A list produced by the function \emph{PrepareInputSequences} or \emph{DistanceMatrix}}
}
\description{
It takes as inputs two pollen sequences produced by the function \emph{PrepareInputSequences}, and computes the best slotting.
}
\examples{
#generating random input data
results.table=GenerateResultsTable(10)
str(results.table)
}
\author{
Blas Benito <blasbenito@gmail.com>
}
