% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateRandomWalk.R
\name{GenerateRandomWalk}
\alias{GenerateRandomWalk}
\title{Generates a random walk.}
\usage{

}
\value{
Data frame, the input random walk with a new column named \emph{distance}
}
\description{
Generates a random walk based on a distance matrix.
Work in progres, put marked code in another function.
}
\author{
Blas Benito <blasbenito@gmail.com>
}

