% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SillyFunctions.R
\name{\%not-in\%}
\alias{\%not-in\%}
\title{Small silly functions}
\usage{
x \%not-in\% y
}
\arguments{
\item{rows}{Number of rows of the results table.}
}
\value{
A dataframe with n rows, and a predefined set of columns
}
\description{
Small silly functions
}
\author{
Blas Benito <blasbenito@gmail.com>
}
