% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HellingerDistance.R
\name{HellingerDistance}
\alias{HellingerDistance}
\title{Hellinger distance between two vectors}
\usage{

}
\arguments{
\item{x}{Numeric vector.}

\item{y}{Numeric vector.}
}
\value{
Numeric, a Hellinger distance
}
\description{
Computes the Hellinger distance between two numeric vectors.
}
\examples{
x=c(1,3,4,5)
y=c(1,3,4,5)
HellingerDistance(x, y)
y=c(23, 45, 67, 24)
HellingerDistance(x, y)
}
\author{
Blas Benito <blasbenito@gmail.com>
}

