% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeastCost.R
\name{LeastCost}
\alias{LeastCost}
\title{Sequence Slotting}
\usage{

}
\arguments{
\item{cost}{A distance matrix produced by \emph{DistanceMatrix}.}

\item{diagonal}{Boolean. If FALSE, the cost is computed}
}
\description{
It takes as inputs two pollen sequences produced by the function \emph{PrepareInputSequences}, and computes the best slotting.
}
\examples{
#generating random input data
results.table=GenerateResultsTable(10)
str(results.table)
}
\author{
Blas Benito <blasbenito@gmail.com>
}
