% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeastCostNNRandom.R
\name{LeastCostNNRandom}
\alias{LeastCostNNRandom}
\title{Search minimum cost path using nearest neighbour.}
\usage{

}
\arguments{
\item{distance.matrix}{A distance matrix provided by \emph{DistanceMatrix}.}
}
\description{
description
}
\examples{

#generating random input data
results.table=GenerateResultsTable(10)
str(results.table)
}
\author{
Blas Benito <blasbenito@gmail.com>
}

