% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManhattanDistance.R
\name{ManhattanDistance}
\alias{ManhattanDistance}
\title{Manhattan distance between two vectors}
\usage{

}
\arguments{
\item{x}{Numeric vector.}

\item{y}{Numeric vector.}
}
\value{
Numeric, a manhattan distance
}
\description{
Computes the manhattan distance between two numeric vectors.
}
\examples{
x=c(1,3,4,5)
y=c(1,3,4,5)
ManhattanDistance(x, y)
y=c(23, 45, 67, 24)
ManhattanDistance(x, y)
}
\author{
Blas Benito <blasbenito@gmail.com>
}

