% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDistanceMatrix.R
\name{PlotDistanceMatrix}
\alias{PlotDistanceMatrix}
\title{running description}
\usage{

}
\arguments{
\item{param}{Number of rows of the results table.}
}
\value{
whatever
\itemize{
 \item{"parameter 1"}{Stuff}
 \item{"parameter 2"}{Stuff}
}
}
\description{
description
}
\examples{

#generating random input data
results.table=GenerateResultsTable(10)
str(results.table)
}
\author{
Blas Benito <blasbenito@gmail.com>
}

