% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSequenceSlotting.R
\name{PlotSequenceSlotting}
\alias{PlotSequenceSlotting}
\title{running description}
\usage{

}
\arguments{
\item{param}{Number of rows of the results table.}
}
\value{
whatever
\itemize{
 \item{"distance.matrix"}{Matrix of distance among samples of both sequences. It is a rectangular matrix}
 \item{"title"}{String. Plot title.}
 \item{"path"}{Path where to store the plot.}
}
}
\description{
description
}
\examples{

#generating random input data
results.table=GenerateResultsTable(10)
str(results.table)
}
\author{
Blas Benito <blasbenito@gmail.com>
}
