% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareSequences.R
\name{PrepareSequences}
\alias{PrepareSequences}
\title{Prepare input sequences.}
\usage{
PrepareSequences(sequence.A = NULL, sequence.A.name = NULL,
  sequence.B = NULL, sequence.B.name = NULL, if.empty.cases = NULL,
  transformation = NULL, silent = NULL)
}
\arguments{
\item{sequence.A}{A dataframe containing pollen counts or percentages. This sequence will be used as reference. It will be checked for empty cases.}

\item{sequence.A.name}{A character string with the name of the sequence.}

\item{sequence.B}{A dataframe containing pollen data. This sequence will be compared with and adapted to the structure of sequence.A}

\item{sequence.B.name}{A character string with the name of the sequence.}

\item{if.empty.cases}{A character argument with two possible values "omit", or "zero". The default value is "omit", but it removes every row with at least one empty record. The option "zero" replaces NA data with zeros.}

\item{transformation}{"none", "percentage", "proportion", "hellinger"}

\item{silent}{Boolean, set to TRUE to hide all messages, and set to FALSE otherwise.}
}
\value{
A list with four slots:
\itemize{
\item \emph{taxa} Common column names of the sequences, listing the taxa included in them.
\item \emph{metadata} A dataframe with details about the processing of the sequences.
\item \emph{sequence.A} Dataframe, a processed pollen sequence.
\item \emph{sequence.B} Dataframe, a processed pollen sequence.
}
}
\description{
This function takes a pollen sequence as reference (sequence.A), and a target pollen sequence (sequence.B), checks the integrity of the reference sequence, and tries to adapt the target sequence to the characteristics of the reference sequence: matching column names and number of columns, and no missing data. Missing data can be handled in two ways: 1) deleting rows with NA or empty cases; 2) interpolation of empty data using the average or weighted average of contiguous cases.
}
\examples{
data(InputDataExample)
}
\author{
Blas Benito <blasbenito@gmail.com>
}
