% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RescalePollenSequence.R
\name{RescalePollenSequence}
\alias{RescalePollenSequence}
\title{Scales pollen data}
\usage{
RescalePollenSequence(sequence, columns = NULL, new.min = NULL,
  new.max = NULL)
}
\arguments{
\item{sequence}{A dataframe containing pollen data. This sequence will be compared with and adapted to the structure of sequence.A}

\item{columns}{Optional, character vector of column names. If not provided, the function will try with all columns.}

\item{new.min}{Minimum new value, 0 by default.}

\item{new.max}{Maximum new value, 100 by default.}
}
\value{
The function returns the reformatted target sequence as a data frame.
}
\description{
Scales pollen data between a minimum and a maximum value.
}
\examples{
data(InputDataExample)
}
\author{
Blas Benito <blasbenito@gmail.com>
}

