% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqSlotBruteForceParallel.R
\name{SeqSlotBruteForceParallel}
\alias{SeqSlotBruteForceParallel}
\title{Sequence Slotting by brute force}
\usage{

}
\arguments{
\item{param}{Number of rows of the results table.}
}
\description{
description
}
\examples{

#generating random input data
results.table=GenerateResultsTable(10)
str(results.table)
}
\author{
Blas Benito <blasbenito@gmail.com>
}

