% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqSlotBruteForceTest.R
\name{SeqSlotBruteForceTest}
\alias{SeqSlotBruteForceTest}
\title{Sequence Slotting by brute force}
\usage{

}
\arguments{
\item{sequences}{A list produced by the function \emph{PrepareInputSequences} or \emph{DistanceMatrix}}

\item{compute.p.value}{Logical value (TRUE or FALSE). Set to TRUE to compute a p-value based on a permutation test.}

\item{max.random.threshold}{Number in the interval [0, 1], determining the proportion of the steps defined by chance during the search for the best slotting solution.}

\item{iterations}{Number of iterations (Default value is 10000).}
}
\description{
It takes as inputs two pollen sequences produced by the function \emph{PrepareInputSequences}, and computes the best slotting.
}
\examples{

#generating random input data
results.table=GenerateResultsTable(10)
str(results.table)
}
\author{
Blas Benito <blasbenito@gmail.com>
}

