% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceSlotting.R
\name{SequenceSlotting}
\alias{SequenceSlotting}
\title{Sequence Slotting}
\usage{

}
\arguments{
\item{sequences}{A list produced by the function \emph{PrepareInputSequences} or \emph{DistanceMatrix}}

\item{compute.p.value}{Boolean. Use TRUE to compute p-values via distance matrix randomization.}

\item{method}{Either "manhattan" computed as \emph{sum(abs(x - y))}, "hellinger" computed as \emph{sqrt(1/2 * sum(sqrt(x)-sqrt(y))^2)}, or "euclidean", computed as \emph{sqrt(sum((x - y)^2))}.}

\item{silent}{Boolean. Use TRUE to remove all messages.}

\item{diagonal}{Boolean. TRUE to compute the sequence slotting using diagonals (true shortest path). FALSE to compute the sequence slotting without considering diagonals. Using diagonals produces lower psi values but better alignments between sequences.}
}
\description{
It takes as inputs two pollen sequences produced by the function \emph{PrepareInputSequences}, and computes the best slotting.
}
\examples{
#generating random input data
results.table=GenerateResultsTable(10)
str(results.table)
}
\author{
Blas Benito <blasbenito@gmail.com>
}
