% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceMatrix.R
\name{DistanceMatrix}
\alias{DistanceMatrix}
\title{Computes a distance matrix}
\usage{

}
\arguments{
\item{sequences}{A list produced by the function \emph{PrepareInputSequences}.}

\item{method}{Either "manhattan" (default) or "hellinger".}
}
\value{
A matrix with manhattan or hellinger distances among cases of the sequences contained in the list \emph{sequences}. The sequence sequence.A will define the rows, and sequence.B will define the columns.
}
\description{
Generates a matrix of manhattan or hellinguer distances between the rows of two pollen sequences.
}
\author{
Blas Benito <blasbenito@gmail.com>
}

